move upload files to separate components
